Sources reorganization
