+++ controls/src/main/java/com/aol/mobile/sdk/controls/ImageLoader.java
