+++ controls/src/main/java/com/aol/mobile/sdk/controls/utils/AndroidHandlerTimer.java
