+++ controls/src/main/java/com/aol/mobile/sdk/controls/utils/Timer.java
