+++ controls/src/main/java/com/aol/mobile/sdk/controls/view/VisibilityWrapper.java
